pyproject.toml
[project]
name = "trailcam-pipeline"
version = "0.1.0"
description = "A command-line tool for producing summary reports from camera trap detections"
license = "MIT"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pydantic>=2.12.5",
    "seaborn>=0.13.2",
    "typer>=0.20.1",
]

[project.scripts]
trailcam = "trailcam_pipeline.cli:cli"

[tool.ruff.lint]
# 1. Enable flake8-bugbear (`B`) rules, in addition to the defaults.
select = ["E4", "E7", "E9", "F", "B"]

# 3. Avoid trying to fix flake8-bugbear (`B`) violations.
unfixable = ["B"]

[tool.pytest]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "matplotlib-stubs>=0.3.11",
    "pytest>=9.0.2",
]
